restructure and fmt
