+++ src/mappers/media-mapper_test.ts
@@ -1,0 +1,4 @@
+import { assertEquals, assertNotEquals } from "./../../test_deps.ts";
+import { copyMedia } from "./media-mapper.ts";
+
+// TODO: Add tests...
