+++ src/resolvers/dublin-core-resolver.ts
@@ -1,4 +1,4 @@
-import { DublinCoreFields } from "../types/dublin-core.ts";
+import { DublinCoreFields } from "../types/fields/mod.ts";
 import { ResolverResult } from "./types/resolver-result.ts";
 
 export const resolveDublinCoreField = (
