+++ src/resolvers/dublin-core-resolver_test.ts
@@ -1,5 +1,5 @@
 import { assertEquals, assertNotEquals } from "../../test_deps.ts";
-import { DublinCoreFields } from "../types/dublin-core.ts";
+import { DublinCoreFields } from "../types/fields/mod.ts";
 import { resolveDublinCoreField } from "./dublin-core-resolver.ts";
 
 [
