+++ src/types/atom.ts
@@ -1,7 +1,6 @@
+import { AtomFields } from "./fields/atom-fields.ts";
+
 // https://validator.w3.org/feed/docs/atom.html
-
-import { AtomFields } from "../resolvers/types/atom-fields.ts";
-
 export interface Atom {
   id: string;
   title: Text;
