+++ src/types/fields/slash-fields.ts
@@ -1,0 +1,6 @@
+export enum SlashFields {
+  Department = "slash:department",
+  Section = "slash:section",
+  Comments = "slash:comments",
+  HitParade = "slash:hit_parade",
+}
