+++ src/types/mod.ts
@@ -1,5 +1,5 @@
 export type { RSS1 } from "./rss1.ts";
-export type { RSS2, Rss2Item, Rss2Channel } from "./rss2.ts";
+export type { RSS2, Rss2Channel, Rss2Item } from "./rss2.ts";
 export type { Atom } from "./atom.ts";
 export type { Feed, FeedEntry } from "./feed.ts";
 export type { JsonFeed } from "./json-feed.ts";
@@ -9,3 +9,4 @@ export type { JsonFeedHub } from "./json-feed.ts";
 export type { DeserializationResult } from "./deserialization-result.ts";
 export { FeedType } from "./feed-type.ts";
 export { FeedParseType } from "./feed-type.ts";
+export { DublinCoreFields } from "./fields/dublin-core-fields.ts";
