+++ src/util_test.ts
@@ -1,5 +1,5 @@
 import { assert, assertEquals } from "../test_deps.ts";
-import { isValidURL, copyValueFields } from "./util.ts";
+import { copyValueFields, isValidURL } from "./util.ts";
 
 Deno.test("is valid HTTP URL", () => {
   assert(isValidURL("https://test.test"));
@@ -8,22 +8,19 @@ Deno.test("is valid HTTP URL", () => {
   assert(!isValidURL("test/data/test"));
 });
 
+Deno.test("CopyValueFields", () => {
+  let target = {
+    a: undefined,
+    b: undefined,
+    c: "cValue",
+  };
+  let source = {
+    a: { value: "aValue" },
+    b: "bValue",
+  };
 
-Deno.test('CopyValueFields', () => {
-
-	let target = {
-		a: undefined,
-		b: undefined,
-		c: 'cValue'
-
-	};
-	let source = {
-		a: { value: 'aValue' },
-		b: 'bValue',
-	};
-
-	copyValueFields(['a', 'b', 'c'], source, target);
-	assertEquals(target.a, 'aValue');
-	assertEquals(target.b, 'bValue');
-	assertEquals(target.c, 'cValue');
+  copyValueFields(["a", "b", "c"], source, target);
+  assertEquals(target.a, "aValue");
+  assertEquals(target.b, "bValue");
+  assertEquals(target.c, "cValue");
 });
